feat: Add mood tracking plugin with emoji-based quick entry
- Implement MoodPlugin with 5-point scale (Great to Bad) with emojis
- Create MoodQuickAddDialog with grid selection and optional notes
- Update QuickAddBottomSheet to support mood-specific dialog
- Update DataScreen to properly display mood entries with emoji and notes
- Add DataSummaryCard component for future analytics features
- Register MoodPlugin in PluginRegistry
- Fix LinearProgressIndicator type conversion issue in DataSummaryCard

The app now supports both water intake and mood tracking through the
plugin system, demonstrating the extensibility of the architecture.

+++ app/src/main/java/com/domain/app/ui/components/DataSummaryCard.kt
@@ -57,7 +57,7 @@ fun WaterSummary(dataPoints: List<DataPoint>) {
     Column {
         Text("Today: ${todayTotal.toInt()} ml")
         LinearProgressIndicator(
-            progress = (todayTotal / 2000f).coerceIn(0f, 1f),
+            progress = (todayTotal / 2000.0).toFloat().coerceIn(0f, 1f),
             modifier = Modifier.fillMaxWidth()
         )
         Text(
